% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_hire_osm.R
\docType{data}
\name{cycle_hire_osm}
\alias{cycle_hire_osm}
\title{TITLE}
\format{FORMAT:
\itemize{
    \item{osm_id} {}
    \item{name} {}
    \item{capacity} {}
    \item{cyclestreets_id} {}
    \item{description} {}
    \item{geometry} {sfc_POINT}
}}
\source{
\url{http://SOURCE.com}
}
\description{
DESCRIPTION
}
\examples{
library(sf)
data(cycle_hire_osm)
# or
cycle_hire_osm <- st_read(system.file("shapes/cycle_hire_osm.geojson", package="spData"))

plot(cycle_hire_osm)

}
\keyword{datasets}
\keyword{sf}
