% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev.R
\docType{data}
\name{elev}
\alias{elev}
\title{Artificial elevation raster data set}
\format{An object of class \code{RasterLayer} of dimension 6 x 6 x 1.}
\usage{
elev
}
\description{
The raster data represents elevation in meters and uses WGS84 as
  a coordinate reference system.
}
\examples{
library(raster)
data(elev)

}
\keyword{datasets}
\keyword{raster}
