% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev.R
\docType{data}
\name{elev.tif}
\alias{elev.tif}
\title{Artificial elevation raster data set}
\description{
The raster data represents elevation in meters and uses WGS84 as
  a coordinate reference system.
}
\examples{
system.file("raster/elev.tif", package = "spData")
}
\keyword{datasets}
\keyword{raster}
