% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grain.R
\docType{data}
\name{grain.tif}
\alias{grain.tif}
\title{Artificial raster dataset representing grain sizes}
\description{
The ratified raster dataset represents grain sizes with the
  three classes clay, silt and sand, and WGS84 as a coordinate reference
  system.
}
\examples{
system.file("raster/grain.tif", package = "spData")
}
\keyword{datasets}
\keyword{raster}
