% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huddersfield.R
\docType{data}
\name{huddersfield}
\alias{huddersfield}
\title{Prevalence of respiratory symptoms}
\format{A data frame with 71 observations on the following 2 variables.
\itemize{
        \item{\code{cases}} {Prevalence of at least mild conditions}
        \item{\code{total}} {Number of questionnaires returned}
}}
\source{
Martuzzi M, Elliott P (1996) Empirical Bayes estimation of small area prevalence of non-rare conditions, Statistics in Medicine 15, 1867--1873, pp. 1870--1871.
}
\description{
Prevalence of respiratory symptoms in 71 school catchment areas in Huddersfield, Northern England
}
\examples{
data(huddersfield)
str(huddersfield)
}
\keyword{datasets}
\keyword{misc}

