% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_hire.R
\docType{data}
\name{cycle_hire}
\alias{cycle_hire}
\title{TITLE}
\format{FORMAT:
\itemize{
    \item{id} {}
    \item{name} {}
    \item{area} {}
    \item{nbikes} {}
    \item{nempty} {}
    \item{geometry} {sfc_POINT}
}}
\source{
\url{http://SOURCE.com}
}
\description{
DESCRIPTION
}
\examples{
library(sf)
data(cycle_hire)
# or
cycle_hire <- st_read(system.file("shapes/cycle_hire.geojson", package="spData"))

plot(cycle_hire)

}
\keyword{datasets}
\keyword{sf}
