% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grain.R
\docType{data}
\name{grain}
\alias{grain}
\title{Artificial raster dataset representing grain sizes}
\format{An object of class \code{RasterLayer} of dimension 6 x 6 x 1.}
\usage{
grain
}
\description{
The ratified raster dataset represents grain sizes with the
  three classes clay, silt and sand, and WGS84 as a coordinate reference
  system.
}
\examples{
if (requireNamespace("raster", quietly = TRUE)) {
  library(raster)
  data(grain)
  grain
}
}
\keyword{datasets}
\keyword{raster}
