% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnd.R
\docType{data}
\name{lnd}
\alias{lnd}
\title{TITLE}
\format{FORMAT:
\itemize{
    \item{NAME} {}
    \item{GSS_CODE} {}
    \item{HECTARES} {}
    \item{NONLD_AREA} {}
    \item{ONS_INNER} {}
    \item{SUB_2009} {}
    \item{SUB_2006} {}
    \item{geometry} {sfc_MULTIPOLYGON}
}}
\source{
\url{http://SOURCE.com}
}
\description{
DESCRIPTION
}
\examples{
library(sf)
data(lnd)
# or
lnd <- st_read(system.file("shapes/lnd.geojson", package="spData"))

plot(lnd)

}
\keyword{datasets}
\keyword{sf}

