% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplashDams.R
\docType{data}
\name{SplashDams}
\alias{SplashDams}
\title{Data for Splash Dams in western Oregon}
\format{Formal class 'SpatialPointsDataFrame with 232 obs. of  6 variables:
\itemize{
\item{streamName} {}
\item{locationCode} {}
\item{height} {}
\item{lastDate} {}
\item{owner} {}
\item{datesUsed} {}
}}
\source{
R. R. Miller (2010) Is the Past Present? Historical Splash-dam Mapping and Stream Disturbance Detection in the Oregon Coastal Province. MSc. thesis, Oregon State University; packaged by Jonathan Callahan
}
\description{
Data for Splash Dams in western Oregon
}
\examples{
data(SplashDams)
library(sp)
plot(SplashDams, axes=TRUE)
}
\references{
\url{http://www.fs.fed.us/pnw/lwm/aem/docs/burnett/miller_rebecca_r2010rev.pdf}
}
\keyword{datasets}

